+++ pkg/bievents/event_test.go
@@ -2,12 +2,16 @@ package bievents
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
+	"fmt"
 	"reflect"
 	"strconv"
 	"strings"
 	"testing"
 
+	"github.com/gesundheitscloud/go-svc/pkg/jwt"
+	"github.com/gofrs/uuid"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -51,6 +55,7 @@ func TestLog(t *testing.T) {
 				"user-id":              "def",
 				"consent-document-key": "consent",
 				"event-source":         "",
+				"session-id":           "",
 				"data": map[string]interface{}{
 					"cuc-id":       "cuc_1",
 					"account-type": "internal",
@@ -81,6 +86,7 @@ func TestLog(t *testing.T) {
 				"user-id":              "def",
 				"consent-document-key": "",
 				"event-source":         "",
+				"session-id":           "",
 				"data": map[string]interface{}{
 					"cuc-id":       "cuc_2",
 					"account-type": "internal",
@@ -107,6 +113,9 @@ func TestLog(t *testing.T) {
 			delete(res, "timestamp")
 			delete(res, "event-id")
 
+			assert.True(t, strings.Contains(fmt.Sprintf("%s", res["event-source"]), "pkg/bievents/event_test.go"))
+			res["event-source"] = ""
+
 			if want, have := tc.result, res; !reflect.DeepEqual(want, have) {
 				t.Errorf("expected values to be %q, got %q", want, have)
 			}
@@ -116,9 +125,37 @@ func TestLog(t *testing.T) {
 }
 
 func TestGetEventSource(t *testing.T) {
-	eventSoource := GetEventSource(0)
-	parts := strings.Split(eventSoource, ":")
+	eventSource := GetEventSource(0)
+	parts := strings.Split(eventSource, ":")
 	assert.True(t, strings.HasSuffix(parts[0], "pkg/bievents/event.go"))
 	_, err := strconv.Atoi(parts[1])
 	assert.NoError(t, err)
+}
+
+func Test_LogCtx(t *testing.T) {
+	buf := new(bytes.Buffer)
+	serviceName := "vega"
+	version := "v1.0.0"
+	hostname := "vega-123-123"
+	userID := uuid.Must(uuid.NewV4())
+	appID := uuid.Must(uuid.NewV4())
+
+	e := NewEventEmitter(serviceName, version, hostname, WithWriter(buf))
+
+	err := e.LogCtx(jwt.NewContext(context.Background(), &jwt.Claims{UserID: userID, AppID: appID}), Event{})
+	assert.NoError(t, err)
+
+	var res map[string]interface{}
+	err = json.NewDecoder(buf).Decode(&res)
+	assert.NoError(t, err)
+
+	assert.True(t, strings.Contains(fmt.Sprintf("%s", res["event-source"]), "pkg/bievents/event_test.go"))
+
+	if v, ok := res["session-id"]; ok {
+		assert.Equal(t, 64, len(fmt.Sprintf("%s", v)))
+	}
+
+	if v, ok := res["user-id"]; ok {
+		assert.Equal(t, userID.String(), fmt.Sprintf("%s", v))
+	}
 }